backend/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-extensions==3.2.3
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
inflection==0.5.1
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
Pillow==10.2.0 # Adding item images to db
geopy==2.4.1 # calculating distance between houses